+++ frontend/src/features/tasks/components/TaskList.tsx
@@ -1,12 +1,12 @@
-import "../components/TaskList.css"
+import "../../../components/TaskList.css"
 import { Pencil, Trash2, LogOut } from "lucide-react";
 import { useEffect, useState, useCallback } from "react";
-import type { Task } from "../types/task";
-import { getMe, fetchTasks, toggleTaskComplete, deleteTask } from "../lib/api";
-import { TaskAdd } from "../components/TaskAdd";
-import EditTaskModal from "../components/EditTaskModal"
+import type { Task } from "../../../types/task";
+import { getMe, fetchTasks, toggleTaskComplete, deleteTask } from "../../../lib/api";
+import { TaskAdd } from "../../../components/TaskAdd";
+import EditTaskModal from "../../../components/EditTaskModal"
 import { Navigate, useNavigate } from "react-router-dom"
-import { useApiError } from "../hooks/useApiError";
+import { useApiError } from "../../../hooks/useApiError";
 
 type User = {
   id: number;
@@ -21,9 +21,6 @@ const getTodayString = () => {
   const dd = String(today.getDate()).padStart(2, "0");
   return `${yyyy}-${mm}-${dd}`;
 };
-
-
-
 
 // 今日のタスク一覧
 export const TaskList = () => {
